chore: move event editor form to its own class, EventEditorForm
